Revert "Feat: Separete ContractService in SAC and smart contract" (#104)
Reverts keizai-tools/keizai-api#102
