+++ src/common/application/mapper/contract.mapper.ts
@@ -1,5 +1,7 @@
 import { nativeToScVal, scValToNative, xdr } from '@stellar/stellar-sdk';
 
+import { Method } from '@/modules/method/domain/method.domain';
+
 import {
   ContractErrorResponse,
   EncodeEvent,
@@ -13,6 +15,27 @@ import {
 } from '../types/soroban.enum';
 
 export class StellarMapper {
+  SCSpecTypeMap = {
+    SC_SPEC_TYPE_BOOL: 'b',
+    SC_SPEC_TYPE_ERROR: 'error',
+    SC_SPEC_TYPE_U32: 'u32',
+    SC_SPEC_TYPE_I32: 'i32',
+    SC_SPEC_TYPE_U64: 'u64',
+    SC_SPEC_TYPE_I64: 'i64',
+    SC_SPEC_TYPE_TIMEPOINT: 'timepoint',
+    SC_SPEC_TYPE_DURATION: 'duration',
+    SC_SPEC_TYPE_U128: 'u128',
+    SC_SPEC_TYPE_I128: 'i128',
+    SC_SPEC_TYPE_U256: 'u256',
+    SC_SPEC_TYPE_I256: 'i256',
+    SC_SPEC_TYPE_BYTES: 'bytes',
+    SC_SPEC_TYPE_STRING: 'str',
+    SC_SPEC_TYPE_SYMBOL: 'sym',
+    SC_SPEC_TYPE_ADDRESS: 'address',
+    SC_SPEC_TYPE_VEC: 'vec',
+    SC_SPEC_TYPE_MAP: 'map',
+  };
+
   encodeEventToDisplayEvent(events: EncodeEvent[]): EventResponse[] {
     return events.map((event) => {
       return {
@@ -31,8 +54,24 @@ export class StellarMapper {
     });
   }
 
-  getNativeToScVal(param: Param) {
-    return nativeToScVal(param.value, { type: param.type });
+  getScValFromStellarAssetContract(
+    selectedMethod: Partial<Method>,
+  ): xdr.ScVal[] {
+    const params = this.getParamsFromStellarAssetContract(selectedMethod);
+    return params.map((param) =>
+      nativeToScVal(param.value, { type: param.type }),
+    );
+  }
+
+  getParamsFromStellarAssetContract(selectedMethod: Partial<Method>): Param[] {
+    return selectedMethod.params.map((param) => {
+      return {
+        value: param.value,
+        type: this.SCSpecTypeMap[
+          selectedMethod.inputs.find((input) => input.name === param.name).type
+        ],
+      };
+    });
   }
 
   fromScValToDisplayValue(value: xdr.ScVal) {
