+++ src/common/common.module.ts
@@ -4,9 +4,7 @@ import { MethodModule } from '@/modules/method/method.module';
 
 import { StellarMapper } from './application/mapper/contract.mapper';
 import { CONTRACT_SERVICE } from './application/repository/contract.interface.service';
-import { ContractService } from './application/service/contract.service';
-import { StellarAssetContractService } from './application/service/sac-contract.service';
-import { SmartContractService } from './application/service/smart-contract.service';
+import { StellarService } from './application/service/stellar.service';
 import { StellarAdapter } from './infrastructure/stellar/stellar.adapter';
 
 @Module({
@@ -14,17 +12,15 @@ import { StellarAdapter } from './infrastructure/stellar/stellar.adapter';
   providers: [
     {
       provide: CONTRACT_SERVICE,
-      useClass: ContractService,
+      useClass: StellarService,
     },
-    StellarAssetContractService,
-    SmartContractService,
     StellarAdapter,
     StellarMapper,
   ],
   exports: [
     {
       provide: CONTRACT_SERVICE,
-      useClass: ContractService,
+      useClass: StellarService,
     },
   ],
 })
